muevo el fetch de metricas a main
