+++ docker-compose.yml
@@ -40,6 +40,7 @@ services:
 
 volumes:
   pgdata:
+    name: db_education
 
 networks:
   my-network:
