+++ src/controllers/metrics.controller.ts
@@ -21,37 +21,17 @@ export const responseTimeHistogram = new client.Histogram({
 });
 register.registerMetric(responseTimeHistogram);
 
-const cpuUsageGauge = new client.Gauge({
+export const cpuUsageGauge = new client.Gauge({
   name: 'cpu_usage_percent',
   help: 'CPU usage percent',
 });
 register.registerMetric(cpuUsageGauge);
 
-const memoryUsageGauge = new client.Gauge({
+export const memoryUsageGauge = new client.Gauge({
   name: 'memory_usage_bytes',
   help: 'Memory usage of the Node.js process in bytes',
 });
 register.registerMetric(memoryUsageGauge);
-
-// Resource fetch
-let prevCpuUsage = process.cpuUsage();
-let prevHrTime = process.hrtime();
-setInterval(() => {
-  const currCpuUsage = process.cpuUsage(prevCpuUsage);
-  const currHrTime = process.hrtime(prevHrTime);
-
-  prevCpuUsage = process.cpuUsage();
-  prevHrTime = process.hrtime();
-
-  const elapsedHrTimeSeconds = currHrTime[0] + currHrTime[1] / 1e9;
-  const totalCpuMicros = currCpuUsage.user + currCpuUsage.system;
-
-  const cpuPercent = (totalCpuMicros / 1e6 / elapsedHrTimeSeconds) * 100;
-  cpuUsageGauge.set(cpuPercent);
-
-  const memoryUsage = process.memoryUsage();
-  memoryUsageGauge.set(memoryUsage.rss);
-}, 5000);
 
 @Controller('education')
 export class MetricsController {
